Fix ref. impl. data types to allow large graphs
These changes do not seem to affect the performance as the fields
whose data types were changed are not part of the performance
bottlenecks.

+++ bk/graph.cpp
@@ -13,7 +13,7 @@
 #define ORPHAN   ( (arc *) 2 )		/* orphan */
 
 template <typename captype, typename tcaptype, typename flowtype> 
-	Graph<captype, tcaptype, flowtype>::Graph(int node_num_max, int edge_num_max, void (*err_function)(const char *))
+	Graph<captype, tcaptype, flowtype>::Graph(size_t node_num_max, size_t edge_num_max, void (*err_function)(const char *))
 	: node_num(0),
 	  nodeptr_block(NULL),
 	  error_function(err_function)
